examples for last features
git-svn-id: http://multirequest.googlecode.com/svn/trunk@14 e4eb5e9b-18bf-bf1a-73b8-3fb6457b787b

+++ example/index.php
@@ -2,7 +2,9 @@
 
 require_once ('config.php');
 
-$urls = array ('http://forums.somethingawful.com/', 'http://asdlksda.sas', 'http://www.somethingpositive.net/', 'http://www.somethingawful.com/', 'http://awesome-hd.net/', 'http://www.istartedsomething.com/', 'http://www.somewhere.fr/', 'http://forums.tkasomething.com/', 'http://www.somewhereinblog.net/', 'http://www.killsometime.com/', 'http://v.sometrics.com/', 'http://www.fearsome-oekaki.com/', 'http://www.dosomething.org/', 'http://www.avonandsomerset.police.uk/');
+/***************************************************************
+  DEBUG METHODS
+ **************************************************************/
 
 function debug($message) {
 	echo $message . '<br />';
@@ -10,14 +12,14 @@ function debug($message) {
 }
 
 function debugRequestComplete(MultiRequest_Request $request, MultiRequest_Handler $handler) {
-	debug('Request complete: ' . $request->getUrl().' Code: '.$request->getCode().' Time: '.$request->getTime());
+	debug('Request complete: ' . $request->getUrl() . ' Code: ' . $request->getCode() . ' Time: ' . $request->getTime());
 	debug('Requests in waiting queue: ' . $handler->getRequestsInQueueCount());
 	debug('Active requests: ' . $handler->getActiveRequestsCount());
 }
 
 function saveCompleteRequestToFile(MultiRequest_Request $request, MultiRequest_Handler $handler) {
 	$filename = preg_replace('/[^\w\.]/', '', $request->getUrl());
-	file_put_contents(DOWNLOADS_DIR.DIRECTORY_SEPARATOR.$filename, $request->getContent());
+	file_put_contents(DOWNLOADS_DIR . DIRECTORY_SEPARATOR . $filename, $request->getContent());
 }
 
 function prepareDownloadsDir() {
@@ -26,20 +28,37 @@ function prepareDownloadsDir() {
 	$dirIterator = new RecursiveDirectoryIterator($dirPath);
 	$recursiveIterator = new RecursiveIteratorIterator($dirIterator);
 	foreach($recursiveIterator as $path) {
-		if($path->isFile()) {
+		if($path->isFile() && strpos($path->getFilename(), '.')) {
 			unlink($path->getPathname());
 		}
 	}
 }
-
 prepareDownloadsDir(DOWNLOADS_DIR);
+
+/***************************************************************
+  MULTIREQUEST INIT
+ **************************************************************/
 
 $mrHandler = new MultiRequest_Handler();
 $mrHandler->setConnectionsLimit(CONNECTIONS_LIMIT);
-$mrHandler->setTimeLimit(TIME_LIMIT);
-$mrHandler->addOnRequestCompleteCallback('debugRequestComplete');
-$mrHandler->addOnRequestCompleteCallback('saveCompleteRequestToFile');
+$mrHandler->onRequestComplete('debugRequestComplete');
+$mrHandler->onRequestComplete('saveCompleteRequestToFile');
 
+$headers = array();
+$headers[] = 'Accept: text/xml,application/xml,application/xhtml+xml,text/html;q=0.9,text/plain;q=0.8,image/png,*/*;q=0.5';
+$headers[] = 'Cache-Control: no-cache';
+$headers[] = 'Connection: Keep-Alive';
+$headers[] = 'Keep-Alive: 300';
+$headers[] = 'Accept-Charset: UTF-8,Windows-1251,ISO-8859-1;q=0.7,*;q=0.7';
+$headers[] = 'Accept-Language: ru,en-us,en;q=0.5';
+$headers[] = 'Pragma:';
+$mrHandler->requestsDefaults()->addHeaders($headers);
+
+$options = array();
+$options[CURLOPT_USERAGENT] = 'Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US; rv:1.9.2.12) Gecko/20101026 Firefox/3.6.12';
+$mrHandler->requestsDefaults()->addCurlOptions($options);
+
+$urls = array('http://forums.somethingawful.com/', 'http://asdlksda.sas', 'http://www.somethingpositive.net/', 'http://www.somethingawful.com/', 'http://awesome-hd.net/', 'http://www.istartedsomething.com/', 'http://www.somewhere.fr/', 'http://forums.tkasomething.com/', 'http://www.somewhereinblog.net/', 'http://www.killsometime.com/', 'http://v.sometrics.com/', 'http://www.fearsome-oekaki.com/', 'http://www.dosomething.org/', 'http://www.avonandsomerset.police.uk/');
 foreach($urls as $url) {
 	$request = new MultiRequest_Request($url);
 	$mrHandler->pushRequestToQueue($request);
@@ -47,6 +66,7 @@ foreach($urls as $url) {
 
 $startTime = time();
 
-$mrHandler->exec();
+set_time_limit(300);
+$mrHandler->start();
 
-debug('Total time: '.(time() - $startTime));
+debug('Total time: ' . (time() - $startTime));
